requirements.txt
discord
pushbullet.py
python-dotenv